# src/backend/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Bicep Tutorial API
  version: 1.0.0
  description: API documentation for the Bicep Tutorial backend.
servers:
  - url: http://localhost:3000
paths:
  /cars:
    get:
      summary: Retrieve a list of cars
      tags:
        - Cars
      responses:
        '200':
          description: A list of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Car'
    post:
      summary: Add a new car
      tags:
        - Cars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCar'
      responses:
        '201':
          description: Car created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cars/{id}:
    get:
      summary: Retrieve a specific car by ID
      tags:
        - Cars
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the car to retrieve
      responses:
        '200':
          description: A single car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a specific car by ID
      tags:
        - Cars
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the car to delete
      responses:
        '200':
          description: Car deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Car with ID 1 deleted successfully.
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a specific car by ID
      tags:
        - Cars
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the car to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCar'
      responses:
        '200':
          description: Car updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Partially update a specific car by ID
      tags:
        - Cars
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the car to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCar'
      responses:
        '200':
          description: Car updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          example: 1
        make:
          type: string
          example: Toyota
        model:
          type: string
          example: Camry
        year:
          type: integer
          example: 2020
    NewCar:
      type: object
      properties:
        make:
          type: string
          example: Honda
        model:
          type: string
          example: Accord
        year:
          type: integer
          example: 2019
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid car data provided.